requirements.txt
pwn
python-magic

